Initial presss of down arrow now increments
Initial selection is index -1 to indicate no selection, which by default
implies the first selection.  The initial press of the down arrow did
not appear to the user as doing anything (and really had no effect),
this changes the behavior so that the initial press skips index 0, and
goes straight to 1 which is probably what the user is expecting.